+++ src/commons-only-untested.spec.ts
@@ -1,9 +1,9 @@
 // import untested service for code-coverage
-import {StringUtils} from "./commons/utils/string.utils";
-import {CommonSqlKeywordAdapter} from "./action-commons/actions/common-sql-keyword.adapter";
-import {CommonSqlObjectDetectionAdapter} from "./action-commons/actions/common-sql-object-detection.adapter";
-import {CommonSqlPlaylistAdapter} from "./action-commons/actions/common-sql-playlist.adapter";
-import {CommonSqlRateAdapter} from "./action-commons/actions/common-sql-rate.adapter";
+import {StringUtils} from './commons/utils/string.utils';
+import {CommonSqlKeywordAdapter} from './action-commons/actions/common-sql-keyword.adapter';
+import {CommonSqlObjectDetectionAdapter} from './action-commons/actions/common-sql-object-detection.adapter';
+import {CommonSqlPlaylistAdapter} from './action-commons/actions/common-sql-playlist.adapter';
+import {CommonSqlRateAdapter} from './action-commons/actions/common-sql-rate.adapter';
 
 for (const a in [
     StringUtils,